feature/dev-GonzaM (backend/offer) fixed price in model and controller

+++ src/database/model/offer.model.js
@@ -24,7 +24,7 @@ const Offer = sequelize.define(
             type: DataTypes.INTEGER()
         },
         price: {
-            type: DataTypes.INTEGER()
+            type: DataTypes.FLOAT()
         }
     }
 );
